git-svn-id: file:///srv/local/data/cleanRepo/trunk@19503 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/yu/src/main/java/playground/yu/integration/cadyts/parameterCalibration/withCarCounts/testLeftTurnWithUnitiveStdDev/PCCtlListener.java
@@ -817,8 +817,8 @@ public class PCCtlListener extends BseParamCalibrationControlerListener
 			String maxParameterNormChangeStr = config.findParam(
 					BSE_CONFIG_MODULE_NAME, "maxParameterNormChange");
 			if (maxParameterNormChangeStr != null) {
-				int maxParameterNormChange = Integer
-						.parseInt(maxParameterNormChangeStr);
+				double maxParameterNormChange = Double
+						.parseDouble(maxParameterNormChangeStr);
 				calibrator.setMaxParameterNormChange(maxParameterNormChange);
 				System.out.println("BSE:\tmaxParameterNormChange\t="
 						+ maxParameterNormChange);
